Virtual keyboard integration behind feature flag `virtual-input`
