+++ wooting-analog-sdk/src/plugin.rs
@@ -5,16 +5,16 @@ use hidapi::DeviceInfo as DeviceInfoHID;
 use hidapi::{HidApi, HidDevice};
 use log::*;
 use log::{error, info};
-use shared_memory::ShmemConf;
 use std::borrow::Borrow;
 use std::collections::HashMap;
 use std::os::raw::{c_float, c_ushort};
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::{Arc, Mutex};
 use std::thread::JoinHandle;
-use std::time::Duration;
 use std::{str, thread};
 
+#[cfg(feature = "virtual-input")]
+use crate::virtual_input::VirtualKeyboard;
 use crate::{DeviceEventType, DeviceID, DeviceInfo, DeviceType, SDKResult, WootingAnalogResult};
 
 #[cfg(target_os = "macos")]
@@ -70,25 +70,6 @@ const ANALOG_BUFFER_SIZE: usize = 48;
 const ANALOG_MAX_SIZE: usize = 40;
 const WOOTING_VID: u16 = 0x31e3;
 const WOOTING_PID_MODE_MASK: u16 = 0xFFF0;
-
-#[derive(Debug, PartialEq)]
-pub struct SharedState {
-    pub vendor_id: u16,
-    /// Device Product ID `pid`
-    pub product_id: u16,
-    //TODO: Consider switching these to FFiStr
-    /// Device Manufacturer name
-    pub manufacturer_name: [u8; 20],
-    /// Device name
-    pub device_name: [u8; 20],
-
-    pub device_type: DeviceType,
-
-    pub device_connected: bool,
-    pub dirty_device_info: bool,
-
-    pub analog_values: [u8; 0xFF],
-}
 
 /// Struct holding the information we need to find the device and the analog interface
 struct DeviceHardwareID {
@@ -348,8 +329,8 @@ pub struct WootingPlugin {
     initialised: Arc<AtomicBool>,
     device_event_cb: Arc<Mutex<Option<Box<dyn Fn(DeviceEventType, &DeviceInfo) + Send>>>>,
     devices: Arc<Mutex<HashMap<DeviceID, Device>>>,
-    #[cfg(debug_assertions)]
-    virtual_keyboard_buffer: Arc<Mutex<HashMap<u16, f32>>>,
+    #[cfg(feature = "virtual-input")]
+    virtual_keyboard: VirtualKeyboard,
     thread: Option<JoinHandle<()>>,
 }
 
@@ -360,8 +341,8 @@ impl WootingPlugin {
             initialised: Arc::new(false.into()),
             device_event_cb: Arc::new(Mutex::new(None)),
             devices: Arc::new(Mutex::new(Default::default())),
-            #[cfg(debug_assertions)]
-            virtual_keyboard_buffer: Arc::new(Mutex::new(HashMap::new())),
+            #[cfg(feature = "virtual-input")]
+            virtual_keyboard: VirtualKeyboard::new(),
             thread: None,
         }
     }
@@ -490,145 +471,8 @@ impl WootingPlugin {
             }
         }));
 
-        // TODO: incorporate virtual keyboard in a better way
-        #[cfg(debug_assertions)]
-        {
-            let t_device_event_cb2 = Arc::clone(&self.device_event_cb);
-
-            let device: Arc<Mutex<Option<DeviceInfo>>> = Arc::new(Mutex::new(None));
-            let device_connected: Arc<Mutex<bool>> = Arc::new(Mutex::new(false));
-            let thread_running: Arc<AtomicBool> = Arc::new(AtomicBool::new(true));
-
-            let t_buffer = Arc::clone(&self.virtual_keyboard_buffer);
-            let t_device = Arc::clone(&device);
-            let t_device_connected = Arc::clone(&device_connected);
-            let t_thread_running = Arc::clone(&thread_running);
-
-            thread::spawn(move || {
-                let link_path = std::env::temp_dir().join("wooting-test-plugin.link");
-                let my_shmem = {
-                    match ShmemConf::new()
-                        .size(4096)
-                        .flink(link_path.as_os_str())
-                        .open()
-                    {
-                        Ok(v) => v,
-                        Err(e) => {
-                            if link_path.exists() {
-                                warn!("Error : {}", e);
-                                warn!(
-                                    "Attempted to open exist SharedMemFailed... Falling back to creation"
-                                );
-                                if let Err(e) = std::fs::remove_file(&link_path) {
-                                    error!("Could not delete old link file: {}", e);
-                                }
-                            }
-                            match ShmemConf::new()
-                                .size(4096)
-                                .flink(link_path.as_os_str())
-                                .create()
-                            {
-                                Ok(m) => m,
-                                Err(e) => {
-                                    error!("Test Plugin Error : {}", e);
-                                    error!("Test Plugin Failed to create SharedMem closing!");
-                                    //return;
-                                    return;
-                                }
-                            }
-                        }
-                    }
-                };
-
-                info!("{:?}", my_shmem.get_flink_path());
-
-                {
-                    let shared_state = unsafe { &mut *(my_shmem.as_ptr() as *mut SharedState) };
-                    shared_state.vendor_id = 0x03eb;
-                    shared_state.product_id = 0xFFFF;
-                    shared_state.device_type = DeviceType::Keyboard;
-                    shared_state.device_connected = false;
-                    shared_state.dirty_device_info = false;
-                    let src = b"Wooting\x00";
-                    shared_state.manufacturer_name[0..src.len()].copy_from_slice(src);
-                    let src = b"Test Device\x00";
-                    shared_state.device_name[0..src.len()].copy_from_slice(src);
-                    shared_state.analog_values = [0; 0xFF];
-                }
-
-                let mut vals = vec![0; 0xFF];
-                loop {
-                    if !t_thread_running.load(Ordering::SeqCst) {
-                        break;
-                    }
-
-                    {
-                        let state = unsafe { &mut *(my_shmem.as_ptr() as *mut SharedState) };
-
-                        if state.dirty_device_info || t_device.lock().unwrap().is_none() {
-                            state.dirty_device_info = false;
-                            let dev = DeviceInfo::new_with_id(
-                                state.vendor_id,
-                                state.product_id,
-                                from_ut8f_to_null(
-                                    &state.manufacturer_name[..],
-                                    state.manufacturer_name.len(),
-                                )
-                                .to_string(),
-                                from_ut8f_to_null(&state.device_name[..], state.device_name.len())
-                                    .to_string(),
-                                1,
-                                state.device_type.clone(),
-                            );
-                            t_device.lock().unwrap().replace(dev);
-                        }
-
-                        if *t_device_connected.lock().unwrap() != state.device_connected {
-                            *t_device_connected.lock().unwrap() = state.device_connected;
-                            if let Some(device) = t_device.lock().unwrap().as_ref() {
-                                t_device_event_cb2.lock().unwrap().as_ref().and_then(|cb| {
-                                    cb(
-                                        if state.device_connected {
-                                            DeviceEventType::Connected
-                                        } else {
-                                            DeviceEventType::Disconnected
-                                        },
-                                        device,
-                                    );
-                                    Some(0)
-                                });
-                            }
-                        }
-
-                        if !state.device_connected {
-                            thread::sleep(Duration::from_millis(500));
-                            continue;
-                        }
-
-                        vals.copy_from_slice(&state.analog_values[..]);
-                    }
-
-                    let analog: HashMap<u16, f32> = vals
-                        .iter()
-                        .enumerate()
-                        .filter_map(|(i, &val)| {
-                            if val > 0 {
-                                Some((i as u16, f32::from(val) / 255_f32))
-                            } else {
-                                None
-                            }
-                        })
-                        .collect();
-                    {
-                        let mut m = t_buffer.lock().unwrap();
-                        m.clear();
-                        m.extend(analog);
-                    }
-                    //t_buffer.lock().unwrap().
-                    thread::sleep(Duration::from_millis(10));
-                }
-            });
-        }
+        #[cfg(feature = "virtual-input")]
+        self.virtual_keyboard.attach();
 
         debug!("Started thread");
         Ok(self.devices.lock().unwrap().len() as u32).into()
@@ -660,7 +504,19 @@ impl Plugin for WootingPlugin {
         }
 
         let ret = self.init_worker();
-        self.device_event_cb.lock().unwrap().replace(callback);
+
+        #[cfg(feature = "virtual-input")]
+        {
+            let cb = Arc::new(Mutex::new(Some(callback)));
+            self.device_event_cb = Arc::clone(&cb);
+            self.virtual_keyboard.with_device_events(cb);
+        }
+
+        #[cfg(not(feature = "virtual-input"))]
+        {
+            self.device_event_cb = Arc::new(Mutex::new(Some(callback)));
+        }
+
         self.initialised.store(ret.is_ok(), Ordering::Relaxed);
         ret
     }
@@ -756,13 +612,15 @@ impl Plugin for WootingPlugin {
             if !any_read {
                 Err(error).into()
             } else {
-                #[cfg(debug_assertions)]
-                for (key, value) in self.virtual_keyboard_buffer.lock().unwrap().iter() {
-                    analog
-                        .entry(*key)
-                        .and_modify(|v| *v = v.max(*value))
-                        .or_insert(*value);
-                }
+                #[cfg(feature = "virtual-input")]
+                self.virtual_keyboard.iter_over(|iter| {
+                    for (key, value) in iter {
+                        analog
+                            .entry(*key)
+                            .and_modify(|v| *v = v.max(*value))
+                            .or_insert(*value);
+                    }
+                });
 
                 Ok(analog).into()
             }
@@ -791,14 +649,4 @@ impl Plugin for WootingPlugin {
 
         Ok(devices).into()
     }
-}
-
-fn from_ut8f_to_null(bytes: &[u8], max_len: usize) -> &str {
-    use std::str::from_utf8_unchecked;
-    for i in 0..max_len {
-        if bytes[i] == 0 {
-            return unsafe { from_utf8_unchecked(&bytes[0..i]) };
-        }
-    }
-    panic!("Couldnt find null terminator.");
 }
